llm-observability-sdk/pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Unified auto-instrumentation SDK for LLM observability"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Platform Team" }]
keywords = ["llm", "observability", "tracing", "opentelemetry", "phoenix", "mlflow"]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.optional-dependencies]
phoenix = [
    "openinference-instrumentation-google-adk>=0.1.0",
    "openinference-instrumentation-google-genai>=0.1.0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
]
mlflow = [
    "mlflow>=2.10.0",
]
test = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "google-adk>=1.20.0",
    "google-genai>=1.0.0",
    "pytest>=8.0",
]
all = [
    "llmops[phoenix,mlflow,test]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmops"]

[tool.pytest.ini_options]
testpaths = ["tests"]
